improvement(test_native_backup_benchmark): reuse performance latency calculation decorator
rely on latency_calculator_decorator measurements for reporting stress results to Argus.
