+++ configurations/manager/6TB_dataset.yaml
@@ -26,6 +26,6 @@ append_scylla_yaml:
       port: 443
       https: true
       aws_region: us-east-1
-      iam_role_arn: arn:aws:iam::797456418907:instance-profile/qa-scylla-manager-backup-instance-profile
-
+aws_instance_profile_name_db: 'qa-scylla-manager-backup-instance-profile'
 user_prefix: 'manager-native-backup-benchmark'
+use_hdrhistogram: true
