+++ jenkins-pipelines/manager/manager-backup-native-backup.jenkinsfile
@@ -6,6 +6,6 @@ def lib = library identifier: 'sct@snapshot', retriever: legacySCM(scm)
 managerPipeline(
     backend: 'aws',
     region: 'us-east-1',
-    test_name: 'mgmt_cli_test.ManagerBackupRestoreConcurrentTests.test_native_backup_benchmark',
+    test_name: 'scylla_manager_performance_test.ManagerBackupRestoreConcurrentTests.test_native_backup_benchmark',
     test_config: '''["configurations/manager/6TB_dataset.yaml", "configurations/kms-ear.yaml"]''',
 )
