+++ jenkins-pipelines/manager/ubuntu22-manager-backup-restore-custom-dataset.jenkinsfile
@@ -6,7 +6,7 @@ def lib = library identifier: 'sct@snapshot', retriever: legacySCM(scm)
 managerPipeline(
     backend: 'aws',
     region: 'us-east-1',
-    test_name: 'mgmt_cli_test.ManagerRestoreBenchmarkTests.test_restore_benchmark',
+    test_name: 'scylla_manager_performance_test.ManagerRestoreBenchmarkTests.test_restore_benchmark',
     test_config: 'test-cases/manager/manager-backup-restore-set-dataset.yaml',
     mgmt_restore_extra_params: "--batch-size 2 --parallel 1",
     mgmt_agent_backup_config: "{'checkers': 100, 'transfers': 2, 'low_level_retries': 20}",
