+++ jenkins-pipelines/manager/ubuntu22-manager-restore-data-outside-of-manager.jenkinsfile
@@ -6,7 +6,7 @@ def lib = library identifier: 'sct@snapshot', retriever: legacySCM(scm)
 managerPipeline(
     backend: 'aws',
     region: 'us-east-1',
-    test_name: 'mgmt_cli_test.ManagerRestoreBenchmarkTests.test_restore_data_without_manager',
+    test_name: 'scylla_manager_performance_test.ManagerRestoreBenchmarkTests.test_restore_data_without_manager',
     test_config: 'test-cases/manager/manager-restore-outside.yaml',
 
     backup_bucket_location: 'manager-backup-tests-permanent-snapshots-us-east-1',
