+++ sdcm/argus_results.py
@@ -131,29 +131,12 @@ class ManagerBackupBenchmarkResult(GenericResultTable):
         ]
 
 
-class ReadStressResult(GenericResultTable):
+class ManagerBackupReadResult(GenericResultTable):
     class Meta:
         name = "Read timing"
         description = "Read timing"
         Columns = [
             ColumnMetadata(name="read time", unit="s", type=ResultType.DURATION, higher_is_better=False),
-            ColumnMetadata(name="op rate", unit="op/s", type=ResultType.FLOAT, higher_is_better=True),
-            ColumnMetadata(name="partition rate", unit="pk/s", type=ResultType.FLOAT, higher_is_better=True),
-            ColumnMetadata(name="row rate", unit="row/s", type=ResultType.FLOAT, higher_is_better=True),
-            ColumnMetadata(name="latency 99th percentile", unit="ms", type=ResultType.FLOAT, higher_is_better=False),
-        ]
-
-
-class WriteStressResult(GenericResultTable):
-    class Meta:
-        name = "Write timing"
-        description = "Write timing"
-        Columns = [
-            ColumnMetadata(name="write time", unit="s", type=ResultType.DURATION, higher_is_better=False),
-            ColumnMetadata(name="op rate", unit="op/s", type=ResultType.FLOAT, higher_is_better=True),
-            ColumnMetadata(name="partition rate", unit="pk/s", type=ResultType.FLOAT, higher_is_better=True),
-            ColumnMetadata(name="row rate", unit="row/s", type=ResultType.FLOAT, higher_is_better=True),
-            ColumnMetadata(name="latency 99th percentile", unit="ms", type=ResultType.FLOAT, higher_is_better=False),
         ]
